Extract shared functions and objects, rename constants
