+++ lib/formatters/eslint.js
@@ -1,18 +1,7 @@
-const { underline, dim, reset } = require('chalk');
-const stripAnsi = require('strip-ansi');
+const { dim, reset } = require('chalk');
 const table = require('text-table');
 const { relative } = require('path');
-const { status, summary } = require('./shared');
-
-const tableOptions = Object.freeze({
-  align: ['r', 'l'],
-  stringLength(str) {
-    return stripAnsi(str).length;
-  }
-});
-
-const formatSource = (path, color, icon) =>
-  `${color(icon)}  ${underline(path)}\n`;
+const { formatSource, STATUSES, summary, TABLE_OPTIONS } = require('./shared');
 
 module.exports = function (results, { cwd }) {
   let problemCount = 0;
@@ -24,10 +13,10 @@ module.exports = function (results, { cwd }) {
     const isError = !!errorCount;
     const isWarning = !!warningCount;
     const { color, icon, severity } = isError
-      ? status.error
+      ? STATUSES.error
       : isWarning
-      ? status.warning
-      : status.success;
+      ? STATUSES.warning
+      : STATUSES.success;
 
     if (!isError && !isWarning) {
       successes += formatSource(path, color, icon);
@@ -44,7 +33,7 @@ module.exports = function (results, { cwd }) {
         message,
         dim(ruleId || '')
       ]),
-      tableOptions
+      TABLE_OPTIONS
     )}\n\n`;
   });
