+++ lib/formatters/shared.js
@@ -1,6 +1,7 @@
-const { bold, red, green, yellow } = require('chalk');
+const { bold, red, green, yellow, underline } = require('chalk');
+const stripAnsi = require('strip-ansi');
 
-const status = Object.freeze({
+const STATUSES = Object.freeze({
   error: {
     color: red,
     icon: '\u2716',
@@ -18,6 +19,16 @@ const status = Object.freeze({
   }
 });
 
+const TABLE_OPTIONS = Object.freeze({
+  align: ['r', 'l'],
+  stringLength(str) {
+    return stripAnsi(str).length;
+  }
+});
+
+const formatSource = (path, color, icon) =>
+  `${color(icon)}  ${underline(path)}\n`;
+
 const summary = (linter, problemTotal, sourceTotal) =>
   bold.underline(
     `${linter} found ${problemTotal} ${
@@ -26,6 +37,8 @@ const summary = (linter, problemTotal, sourceTotal) =>
   );
 
 module.exports = {
-  status,
-  summary
+  formatSource,
+  STATUSES,
+  summary,
+  TABLE_OPTIONS
 };
